+++ resources/views/admin/audit/show.blade.php
@@ -1,3 +1,3 @@
-<x-chief::page.template :title="'Audit van ' . $causer->fullname" container="md">
+<x-chief::page.template :title="'Audit van ' . $causerSnapshot['fullname']" container="md">
     @include('chief::admin.audit._rows')
 </x-chief::page.template>